pyproject.toml
[tool.poetry]
name = "secconfig"
version = "0.1.0"
description = ""
authors = ["David (DDT) Tillemans <david.tillemans@cryptable.org>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
secconfig = "secconfig.main:run"